requirements.txt
langchain==0.3.25
langchain-openai==0.3.18
langchain-core==0.3.59
langserve[all]==0.3.1
fastapi==0.115.12
uvicorn==0.34.2
streamlit==1.45.1
python-dotenv==1.1.0
requests==2.32.3
openai==1.82.0
httpx==0.28.1
pydantic==2.11.4
